上架前1.1
后台增加了举报管理，增强了各方面的用户体验；接口接入距离，数据整合规范化，默认头像默认名，增强程序健壮性

+++ Application/Home/Controller/ReturnController.class.php
@@ -6,7 +6,7 @@ class ReturnController extends Controller{
      * 用来返回需要登录权限的接口的预处理的返回信息
      */
     function returnMsg(){
-        $re=I('re');
+        $re=(int)I('re');
         $type=I('type');
         $this->ajaxReturn(responseMsg($re,$type));
     }
